Do not reconstruct workflow on every update
Instead when converting back from node dict, check whether node of the
same label and object id exists and return that.
Edges are still completely rebuild on every update to make sure user
deleted connections really disappear.
This could be improved in a future PR, but seems neither performance nor
feature-wise relevant, so it is not on my radar.

This keeps data caches on nodes and avoids the errors that occured
previously on using a previously saved workflow.